requirements.txt
# Core dependencies for algo-trading system
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Technical indicators
ta>=0.10.2
TA-Lib>=0.4.25

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# Google Sheets integration
gspread>=5.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Telegram bot
python-telegram-bot>=20.3

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
logging>=0.4.9.6
requests>=2.31.0
openpyxl>=3.1.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
jupyter>=1.0.0
